fake_device moved to miio package
